requirements.txt
openai
python-telegram-bot
requests
google-api-python-client
google-analytics-data
gspread
google-auth
google-auth-oauthlib
python-dotenv
edge-tts
chromadb
